deprecate "common" folder, use standard structure
