+++ src/app/app.module.ts
@@ -52,22 +52,22 @@ import { SidenavComponent } from './components/sidenav/sidenav.component';
 import { AboutComponent } from './components/about/about.component';
 import { LocationLinkComponent } from './components/location-link/location-link.component';
 import { SharedLinkViewComponent } from './components/shared-link-view/shared-link-view.component';
-import { NodeCopyDirective } from './common/directives/node-copy.directive';
-import { NodeDeleteDirective } from './common/directives/node-delete.directive';
-import { NodeMoveDirective } from './common/directives/node-move.directive';
-import { NodeRestoreDirective } from './common/directives/node-restore.directive';
-import { NodePermanentDeleteDirective } from './common/directives/node-permanent-delete.directive';
-import { NodeUnshareDirective } from './common/directives/node-unshare.directive';
-import { NodeVersionsDirective } from './common/directives/node-versions.directive';
+import { NodeCopyDirective } from './directives/node-copy.directive';
+import { NodeDeleteDirective } from './directives/node-delete.directive';
+import { NodeMoveDirective } from './directives/node-move.directive';
+import { NodeRestoreDirective } from './directives/node-restore.directive';
+import { NodePermanentDeleteDirective } from './directives/node-permanent-delete.directive';
+import { NodeUnshareDirective } from './directives/node-unshare.directive';
+import { NodeVersionsDirective } from './directives/node-versions.directive';
 import { NodeVersionsDialogComponent } from './dialogs/node-versions/node-versions.dialog';
 import { LibraryDialogComponent } from './dialogs/library/library.dialog';
-import { ContentManagementService } from './common/services/content-management.service';
-import { NodeActionsService } from './common/services/node-actions.service';
-import { NodePermissionService } from './common/services/node-permission.service';
+import { ContentManagementService } from './services/content-management.service';
+import { NodeActionsService } from './services/node-actions.service';
+import { NodePermissionService } from './services/node-permission.service';
 import { SearchResultsComponent } from './components/search/search-results/search-results.component';
 import { SettingsComponent } from './components/settings/settings.component';
-import { ProfileResolver } from './common/services/profile.resolver';
-import { ExperimentalGuard } from './common/services/experimental-guard.service';
+import { ProfileResolver } from './services/profile.resolver';
+import { ExperimentalGuard } from './services/experimental-guard.service';
 
 import { InfoDrawerComponent } from './components/info-drawer/info-drawer.component';
 import { EditFolderDirective } from './directives/edit-folder.directive';
@@ -81,7 +81,7 @@ import { ExtensionsModule } from './extensions.module';
 import { CoreExtensionsModule } from './extensions/core.extensions.module';
 import { SearchResultsRowComponent } from './components/search/search-results-row/search-results-row.component';
 import { NodePermissionsDialogComponent } from './dialogs/node-permissions/node-permissions.dialog';
-import { NodePermissionsDirective } from './common/directives/node-permissions.directive';
+import { NodePermissionsDirective } from './directives/node-permissions.directive';
 import { PermissionsManagerComponent } from './components/permission-manager/permissions-manager.component';
 
 @NgModule({
