+++ src/app/components/files/files.component.ts
@@ -29,8 +29,8 @@ import { ActivatedRoute, Params, Router } from '@angular/router';
 import { Store } from '@ngrx/store';
 import { MinimalNodeEntity, MinimalNodeEntryEntity, NodePaging, PathElement, PathElementEntity } from 'alfresco-js-api';
 import { Observable } from 'rxjs/Rx';
-import { ContentManagementService } from '../../common/services/content-management.service';
-import { NodeActionsService } from '../../common/services/node-actions.service';
+import { ContentManagementService } from '../../services/content-management.service';
+import { NodeActionsService } from '../../services/node-actions.service';
 import { AppStore } from '../../store/states/app.state';
 import { PageComponent } from '../page.component';
 import { ContentApiService } from '../../services/content-api.service';
