+++ src/app/components/shared-files/shared-files.component.ts
@@ -24,7 +24,7 @@
  */
 
 import { Component, OnInit } from '@angular/core';
-import { ContentManagementService } from '../../common/services/content-management.service';
+import { ContentManagementService } from '../../services/content-management.service';
 import { PageComponent } from '../page.component';
 import { Store } from '@ngrx/store';
 import { AppStore } from '../../store/states/app.state';
