+++ src/app/components/sidenav/sidenav.component.ts
@@ -26,7 +26,7 @@
 import { Subject } from 'rxjs/Rx';
 import { Component, Input, OnInit, OnDestroy } from '@angular/core';
 import { Node } from 'alfresco-js-api';
-import { NodePermissionService } from '../../common/services/node-permission.service';
+import { NodePermissionService } from '../../services/node-permission.service';
 import { ExtensionService } from '../../extensions/extension.service';
 import { NavigationExtension } from '../../extensions/navigation.extension';
 import { Store } from '@ngrx/store';
